Load into Google Earth Engine from Google Cloud Storage

Load into Google Earth Engine from Google Cloud Storage

gcs_to_gee(
  gcs_name,
  gee_name = fs::path_ext_remove(gcs_name),
  properties = "",
  gcs_bucket = "offhab_lyrs",
  gee_asset = "projects/ee-offhab/assets/lyrs_aphia_web",
  pyramiding_policy = "MEAN",
  missing_data = 255
)

Arguments

gcs_name

Google Cloud Storage (GCS) file name

gee_name

Google Earth Engine (GEE) file name (cannot have .); defaults to gcs_name without the file extension

properties

a list of properties or blank (default: "")

gcs_bucket

GCS bucket (like a folder) for containing gcs_name; default: "offhab_lyrs"

gee_asset

GEE asset (like a folder) for containing gee_name; default: "projects/ee-offhab/assets/lyrs_aphia_web"

pyramiding_policy

GEE pyramiding; default: "MEAN" (but if categorical data want something else here)

missing_data

GEE missing data; default: 255 (but need something esle if not INT1U)

Value

string of {gee_asset}/{gee_name} to reference image in GEE