Offshore Habitat Zone polygons

Offshore habitat zones used for summary statistics and derived from BOEM 2019-2024 Draft Proposed Program Areas for the lower 48 (excluding Hawaii and Alaska), clipped by the United States Exclusive Economic Zone (EEZ). A second version was further intersected with the OceanAdapt Regions for bottom trawl models, since the available data is markedly different inside versus outside these model results.

oh_zones

Format

A spatial features (sf) data frame with 22 features and 11 fields:

zone_id

unique zone identifier (integer)

zone_version

version (integer) indicating whether with (2) or without (1) OceanAdapt regions

zone_key

unique zone identifier (character), of the form {boem_key}-{oa_key}

zone_name

name of zone, of the form {boem_area}-{oa_region}

region

oceanic BOEM region; one of: "Atlantic", "Gulf of Mexico" or "Pacific"

boem_key

the three letter lower-case code for the BOEM Planning Area, original MMS_PLAN_A column

boem_area

name of BOEM Planning Area, original RESA_summa column

area_km2

area of zone in square kilometers

oa_key

2-letter lowercase key for OceanAdapt region for bottom trawl data

oa_region

name of OceanAdapt region

oa_region_rds

names of regions stored data files (*.rds); multiple values seperated by a ;

geom

geometry in geographic coordinate system (EPSG: 4326)

Source

Geographic Mapping Data in Digital Format, Marine Regions · United States Exclusive Economic Zone (EEZ), DisMAP Regions 20220516 | InPort